requirements.txt
aiogram==2.10.1
aioschedule==0.5.2
asyncio==3.4.3
numpy==1.21.5
gspread==3.6.0
pandas==1.4.1
google-api-python-client==2.0.2
yfinance~=0.1.70
DateTime~=4.3
httplib2~=0.19.0
mibian~=0.1.3
oauth2client~=4.1.3
scipy==1.8.0
pandas_ta==0.3.14b
requests==2.27.1
python-dateutil==2.8.2
ibapi~=9.81.1.post1